requirements.txt
fastapi
uvicorn
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
deprecated
